pyproject.toml
[tool.poetry]
name = "baxxter"
version = "0.1.0"
description = ""
authors = ["karrni <karrni@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-decouple = "^3.8"
yt-dlp = "^2024.4.9"
yarl = "^1.9.4"
discord-py = {version = "^2.3.2", extras = ["voice"]}


[tool.poetry.group.dev.dependencies]
black = "^24.4.1"
isort = "^5.13.2"
ruff = "^0.4.2"


[tool.poetry.group.voice.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
